+++ .env.example
@@ -1,8 +1,5 @@
 CORS_ALLOW_ORIGIN='*'
-DATABASE_URL='postgresql://linx:linx@localhost:5432/linx'
+DATABASE_DIRECTORY='./database'
 HOST='0.0.0.0'
 JWT_SECRET='linx'
-POSTGRES_USER='linx'
-POSTGRES_PASSWORD='linx'
-POSTGRES_DB='linx'
 PORT=3000
