+++ .gitignore
@@ -3,4 +3,4 @@
 
 !.env.example
 .DS_Store
-linx_database.db
+database
