+++ Cargo.toml
@@ -1,6 +1,23 @@
-[workspace]
-members = [
-    "src/entity",
-    "src/migration",
-    "src/server",
-]
+[package]
+name = "linx"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.65"
+rust-argon2 = "1.0.0"
+async-graphql = { version = "4.0.14", features = ["chrono"] }
+async-graphql-axum = "4.0.14"
+axum = "0.5.16"
+chrono = { version = "0.4.22", features = ["serde"] }
+dotenv = "0.15.0"
+jsonwebtoken = "8.0"
+rand = "0.8.5"
+serde = { version = "1.0.144", features = ["derive"] }
+sled = "0.34.7"
+tokio = { version = "1.21.0", features = ["rt", "rt-multi-thread", "macros"] }
+tower-http = { version = "0.3.0", features = ["cors"] }
+tracing = "0.1.37"
+tracing-subscriber = "0.3.15"
+url = "2.3.1"
+uuid = "1.2.2"
