--- docker-compose.yml
@@ -1,14 +1,0 @@
-version: '3'
-
-services:
-  database:
-    image: 'postgres:14.1-alpine'
-    ports:
-      - '5432:5432'
-    volumes:
-      - database:/var/lib/postgresql/data
-    env_file:
-      - .env
-
-volumes:
-  database:
