+++ src/config.rs
@@ -2,12 +2,10 @@ use std::env;
 use std::net::{IpAddr, Ipv4Addr};
 use std::str::FromStr;
 
-use sea_orm::ConnectOptions;
-
 #[derive(Debug)]
 pub struct Config {
     pub cors_allow_origin: String,
-    pub dbconnect_options: ConnectOptions,
+    pub database_path: String,
     pub jwt_secret: String,
     pub server_host: IpAddr,
     pub server_port: u16,
@@ -16,16 +14,15 @@ pub struct Config {
 impl Config {
     pub fn new() -> Self {
         let cors_allow_origin = Config::env_var::<String>("CORS_ALLOW_ORIGIN");
+        let database_path = Config::env_var::<String>("DATABASE_DIRECTORY");
         let jwt_secret = Config::env_var::<String>("JWT_SECRET");
         let server_host = Config::env_var_opt::<IpAddr>("HOST")
             .unwrap_or(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)));
         let server_port = Config::env_var::<u16>("PORT");
-        let database_url = Config::env_var::<String>("DATABASE_URL");
-        let dbconnect_options = ConnectOptions::new(database_url);
 
         Self {
             cors_allow_origin,
-            dbconnect_options,
+            database_path,
             jwt_secret,
             server_host,
             server_port,
