--- src/entity/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "entity"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "entity"
-path = "src/lib.rs"
-
-[dependencies]
-serde = { version = "1", features = ["derive"] }
-sea-orm = { version = "0.9.2" }
