--- src/entity/src/mod.rs
@@ -1,6 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.3
-
-pub mod prelude;
-
-pub mod link;
-pub mod user;
