--- src/entity/src/prelude.rs
@@ -1,4 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.3
-
-pub use super::link::Entity as Link;
-pub use super::user::Entity as User;
