--- src/entity/src/user.rs
@@ -1,32 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.10.3
-
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
-#[sea_orm(table_name = "user")]
-pub struct Model {
-    #[sea_orm(primary_key)]
-    pub id: i32,
-    pub name: String,
-    pub last_name: String,
-    #[sea_orm(unique)]
-    pub email: String,
-    pub hash: String,
-    pub created_at: DateTimeWithTimeZone,
-    pub updated_at: DateTimeWithTimeZone,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(has_many = "super::link::Entity")]
-    Link,
-}
-
-impl Related<super::link::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Link.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
