+++ src/handlers/graphql.rs
@@ -7,7 +7,7 @@ use axum::http::HeaderMap;
 use axum::response::{Html, IntoResponse};
 use axum::Extension;
 
-use crate::graphql::GraphQLSchema;
+// use crate::graphql::GraphQLSchema;
 use crate::services::auth::Token;
 
 pub async fn schema(
