--- src/migration/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "migration"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lib]
-name = "migration"
-path = "src/lib.rs"
-
-[dependencies]
-async-std = { version = "^1", features = ["attributes", "tokio1"] }
-sea-orm-migration = { version = "0.9.2", features = [ "sqlx-postgres", "runtime-tokio-rustls" ] }
