--- src/migration/src/lib.rs
@@ -1,18 +1,0 @@
-mod m20220915_153744_create_links_table;
-mod m20221015_100421_create_users_table;
-mod m20221103_002242_links_table_ownership;
-
-pub use sea_orm_migration::prelude::*;
-
-pub struct Migrator;
-
-#[async_trait::async_trait]
-impl MigratorTrait for Migrator {
-    fn migrations() -> Vec<Box<dyn MigrationTrait>> {
-        vec![
-            Box::new(m20220915_153744_create_links_table::Migration),
-            Box::new(m20221015_100421_create_users_table::Migration),
-            Box::new(m20221103_002242_links_table_ownership::Migration),
-        ]
-    }
-}
