+++ src/models/mod.rs
@@ -1,3 +1,2 @@
 pub mod link;
-pub mod prelude;
 pub mod user;
